Move colors submodule out
