+++ gerrytools/colors/districtr.py
