+++ gerrytools/colors/utils.py
