+++ gerrytools/plotting/__init__.py
@@ -3,16 +3,10 @@ Makes pretty pictures of districting plans, dual graphs, histograms, boxplots,
 and violin plots 🎻.
 """
 
+from gerrytools.colors import districtr, flare, latex, purples, redbluecmap
+
 from .annotation import arrow, ideal
 from .bins import bins
-from .choropleth import choropleth
-from .colors import districtr, flare, latex, purples, redbluecmap
-from .districtnumbers import districtnumbers
-from .drawgraph import drawgraph
-from .drawplan import drawplan
-from .gifs import gif_multidimensional
-from .histogram import histogram
-from .multidimensional import multidimensional
 from .scatterplot import scatterplot
 from .sealevel import sealevel
 from .violin import violin
